refactor(vite-plugin-angular): consolidate tsconfig resolution into TsconfigResolver
Extract the 11 duplicated tsconfig functions (readAngularTsconfigConfiguration,
getCachedTsconfigOptions, collectExpandedTsconfigRoots, collectTsconfigPathRoots,
resolveReferenceTsconfigPath, ensureIncludeCache, findIncludes,
normalizeIncludeGlob, getTsconfigCacheKey, invalidateTsconfigCaches,
invalidateIncludeCache) into a shared TsconfigResolver class in
utils/tsconfig-resolver.ts.

Both angular-vite-plugin.ts and compilation-api-plugin.ts now instantiate
their own TsconfigResolver with independent caches, eliminating ~7k lines
of duplication in each file.
