+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -20,7 +20,6 @@ import * as vite from 'vite';
 import * as compilerCli from '@angular/compiler-cli';
 import { createRequire } from 'node:module';
 import * as ts from 'typescript';
-import { globSync } from 'tinyglobby';
 import {
   defaultClientConditions,
   ModuleNode,
@@ -98,6 +97,10 @@ import {
   TS_EXT_REGEX,
   type TsConfigResolutionContext,
 } from './utils/plugin-config.js';
+import {
+  TsconfigResolver,
+  normalizeIncludeGlob,
+} from './utils/tsconfig-resolver.js';
 import { getJsTransformConfigKey, isRolldown } from './utils/rolldown.js';
 import { toVirtualRawId, toVirtualStyleId } from './utils/virtual-ids.js';
 import { type SourceFileCache as SourceFileCacheType } from './utils/source-file-cache.js';
@@ -291,26 +294,7 @@ export interface PluginOptions {
   };
 }
 
-export function normalizeIncludeGlob(
-  workspaceRoot: string,
-  glob: string,
-): string {
-  const normalizedWorkspaceRoot = normalizePath(resolve(workspaceRoot));
-  const normalizedGlob = normalizePath(glob);
-
-  if (
-    normalizedGlob === normalizedWorkspaceRoot ||
-    normalizedGlob.startsWith(`${normalizedWorkspaceRoot}/`)
-  ) {
-    return normalizedGlob;
-  }
-
-  if (normalizedGlob.startsWith('/')) {
-    return `${normalizedWorkspaceRoot}${normalizedGlob}`;
-  }
-
-  return normalizePath(resolve(normalizedWorkspaceRoot, normalizedGlob));
-}
+export { normalizeIncludeGlob } from './utils/tsconfig-resolver.js';
 
 const classNames = new Map();
 export function evictDeletedFileMetadata(
@@ -382,23 +366,21 @@ export function angular(options?: PluginOptions): Plugin[] {
   const ts = require('typescript');
   let builder: ts.BuilderProgram | ts.EmitAndSemanticDiagnosticsBuilderProgram;
   let nextProgram: NgtscProgram | undefined;
-  // Caches (always rebuild Angular program per user request)
-  const tsconfigOptionsCache = new Map<
-    string,
-    { options: ts.CompilerOptions; rootNames: string[] }
-  >();
-  const tsconfigGraphRootCache = new Map<string, string[]>();
   let cachedHost: ts.CompilerHost | undefined;
   let cachedHostKey: string | undefined;
-  let includeCache: string[] = [];
+  const isTest = process.env['NODE_ENV'] === 'test' || !!process.env['VITEST'];
+  const tsconfigResolver = new TsconfigResolver({
+    workspaceRoot: pluginOptions.workspaceRoot,
+    include: pluginOptions.include,
+    liveReload: pluginOptions.liveReload,
+    hasTailwindCss: pluginOptions.hasTailwindCss,
+    isTest,
+  });
   function invalidateFsCaches() {
-    includeCache = [];
+    tsconfigResolver.invalidateIncludeCache();
   }
   function invalidateTsconfigCaches() {
-    // `readConfiguration` caches the root file list, so hot-added pages can be
-    // missing from Angular's compilation program until we clear this state.
-    tsconfigOptionsCache.clear();
-    tsconfigGraphRootCache.clear();
+    tsconfigResolver.invalidateTsconfigCaches();
     cachedHost = undefined;
     cachedHostKey = undefined;
   }
@@ -467,7 +449,6 @@ export function angular(options?: PluginOptions): Plugin[] {
 
   let stylesheetRegistry: AnalogStylesheetRegistry | undefined;
   const sourceFileCache: SourceFileCacheType = new SourceFileCache();
-  const isTest = process.env['NODE_ENV'] === 'test' || !!process.env['VITEST'];
   const isVitestVscode = !!process.env['VITEST_VSCODE'];
   const isStackBlitz = !!process.versions['webcontainer'];
   const isAstroIntegration = process.env['ANALOG_ASTRO'] === 'true';
@@ -1584,290 +1565,6 @@ export function angular(options?: PluginOptions): Plugin[] {
     encapsulationPlugin(shouldExternalizeStyles),
   ].filter(Boolean) as Plugin[];
 
-  function findIncludes() {
-    // Map include patterns to absolute workspace paths
-    const globs = pluginOptions.include.map((glob) =>
-      normalizeIncludeGlob(pluginOptions.workspaceRoot, glob),
-    );
-
-    // Discover TypeScript files using tinyglobby
-    const files = globSync(globs, {
-      dot: true,
-      absolute: true,
-    });
-
-    debugEmit('include discovery', {
-      patternCount: globs.length,
-      fileCount: files.length,
-    });
-    debugEmitV('include discovery files', {
-      globs,
-      files: files.map((file) => normalizePath(file)),
-    });
-
-    return files;
-  }
-
-  function ensureIncludeCache(): string[] {
-    if (pluginOptions.include.length > 0 && includeCache.length === 0) {
-      includeCache = findIncludes();
-      debugEmit('include cache populated', {
-        fileCount: includeCache.length,
-      });
-    }
-
-    return includeCache;
-  }
-
-  function getTsconfigCacheKey(
-    resolvedTsConfigPath: string,
-    config: ResolvedConfig,
-  ): string {
-    const isProd = config.mode === 'production';
-
-    return [
-      resolvedTsConfigPath,
-      isProd ? 'prod' : 'dev',
-      isTest ? 'test' : 'app',
-      config.build?.lib ? 'lib' : 'nolib',
-      pluginOptions.liveReload ? 'live-reload' : 'no-live-reload',
-      pluginOptions.hasTailwindCss ? 'tw' : 'notw',
-    ].join('|');
-  }
-
-  function readAngularTsconfigConfiguration(
-    resolvedTsConfigPath: string,
-    config: ResolvedConfig,
-  ) {
-    const isProd = config.mode === 'production';
-    return compilerCli.readConfiguration(resolvedTsConfigPath, {
-      suppressOutputPathCheck: true,
-      outDir: undefined,
-      sourceMap: false,
-      inlineSourceMap: !isProd,
-      inlineSources: !isProd,
-      declaration: false,
-      declarationMap: false,
-      allowEmptyCodegenFiles: false,
-      annotationsAs: 'decorators',
-      enableResourceInlining: false,
-      noEmitOnError: false,
-      mapRoot: undefined,
-      sourceRoot: undefined,
-      supportTestBed: false,
-      supportJitMode: false,
-    });
-  }
-
-  function getCachedTsconfigOptions(
-    resolvedTsConfigPath: string,
-    config: ResolvedConfig,
-  ): { options: ts.CompilerOptions; rootNames: string[] } {
-    const tsconfigKey = getTsconfigCacheKey(resolvedTsConfigPath, config);
-    let cached = tsconfigOptionsCache.get(tsconfigKey);
-
-    if (!cached) {
-      const read = readAngularTsconfigConfiguration(
-        resolvedTsConfigPath,
-        config,
-      );
-      cached = { options: read.options, rootNames: read.rootNames };
-      tsconfigOptionsCache.set(tsconfigKey, cached);
-      debugEmit('tsconfig root names loaded', {
-        resolvedTsConfigPath,
-        rootNameCount: read.rootNames.length,
-      });
-      debugEmitV('tsconfig root names', {
-        resolvedTsConfigPath,
-        rootNames: read.rootNames.map((file) => normalizePath(file)),
-      });
-    }
-
-    return cached;
-  }
-
-  function resolveReferenceTsconfigPath(
-    referencePath: string,
-    ownerTsconfigPath: string,
-  ): string | undefined {
-    const ownerDir = dirname(ownerTsconfigPath);
-    const resolvedReference = normalizePath(
-      isAbsolute(referencePath)
-        ? referencePath
-        : resolve(ownerDir, referencePath),
-    );
-
-    if (existsSync(resolvedReference)) {
-      try {
-        if (statSync(resolvedReference).isDirectory()) {
-          const nestedTsconfig = join(resolvedReference, 'tsconfig.json');
-          return existsSync(nestedTsconfig)
-            ? normalizePath(nestedTsconfig)
-            : undefined;
-        }
-      } catch {
-        return undefined;
-      }
-
-      return resolvedReference;
-    }
-
-    if (!resolvedReference.endsWith('.json')) {
-      const asJson = `${resolvedReference}.json`;
-      if (existsSync(asJson)) {
-        return normalizePath(asJson);
-      }
-
-      const nestedTsconfig = join(resolvedReference, 'tsconfig.json');
-      if (existsSync(nestedTsconfig)) {
-        return normalizePath(nestedTsconfig);
-      }
-    }
-
-    return undefined;
-  }
-
-  function collectTsconfigPathRoots(
-    resolvedTsConfigPath: string,
-    options: ts.CompilerOptions,
-    rawTsconfig: {
-      compilerOptions?: {
-        baseUrl?: unknown;
-        paths?: Record<string, string[]>;
-      };
-    },
-  ): string[] {
-    const tsPaths = rawTsconfig.compilerOptions?.paths ?? options.paths;
-    if (!tsPaths) {
-      return [];
-    }
-
-    const tsconfigDir = dirname(resolvedTsConfigPath);
-    const configuredBaseUrl =
-      typeof options.baseUrl === 'string'
-        ? options.baseUrl
-        : typeof rawTsconfig.compilerOptions?.baseUrl === 'string'
-          ? rawTsconfig.compilerOptions.baseUrl
-          : undefined;
-    const resolvedBaseUrl = configuredBaseUrl
-      ? isAbsolute(configuredBaseUrl)
-        ? configuredBaseUrl
-        : resolve(tsconfigDir, configuredBaseUrl)
-      : tsconfigDir;
-    const discoveredRoots = new Set<string>();
-
-    for (const targets of Object.values(tsPaths)) {
-      for (const target of targets) {
-        const resolvedTarget = normalizePath(
-          isAbsolute(target) ? target : resolve(resolvedBaseUrl, target),
-        );
-
-        if (target.includes('*')) {
-          for (const match of globSync(resolvedTarget, {
-            dot: true,
-            absolute: true,
-          })) {
-            discoveredRoots.add(normalizePath(match));
-          }
-          continue;
-        }
-
-        if (existsSync(resolvedTarget)) {
-          discoveredRoots.add(resolvedTarget);
-        }
-      }
-    }
-
-    return [...discoveredRoots];
-  }
-
-  function collectExpandedTsconfigRoots(
-    resolvedTsConfigPath: string,
-    config: ResolvedConfig,
-    visited = new Set<string>(),
-  ): string[] {
-    const normalizedTsConfigPath = normalizePath(resolvedTsConfigPath);
-    if (visited.has(normalizedTsConfigPath)) {
-      return [];
-    }
-
-    const tsconfigKey = `${getTsconfigCacheKey(normalizedTsConfigPath, config)}|graph`;
-    const cached = tsconfigGraphRootCache.get(tsconfigKey);
-    if (cached) {
-      return cached;
-    }
-
-    visited.add(normalizedTsConfigPath);
-
-    const read = readAngularTsconfigConfiguration(
-      normalizedTsConfigPath,
-      config,
-    );
-    const rawTsconfig = (ts.readConfigFile(
-      normalizedTsConfigPath,
-      ts.sys.readFile,
-    ).config ?? {}) as {
-      compilerOptions?: {
-        baseUrl?: unknown;
-        paths?: Record<string, string[]>;
-      };
-      references?: Array<{ path?: unknown }>;
-    };
-
-    const expandedRoots = new Set(
-      read.rootNames.map((file) => normalizePath(file)),
-    );
-    const pathRoots = collectTsconfigPathRoots(
-      normalizedTsConfigPath,
-      read.options,
-      rawTsconfig,
-    );
-    for (const pathRoot of pathRoots) {
-      expandedRoots.add(pathRoot);
-    }
-
-    const referenceConfigs = (rawTsconfig.references ?? [])
-      .flatMap((reference) =>
-        typeof reference.path === 'string'
-          ? [
-              resolveReferenceTsconfigPath(
-                reference.path,
-                normalizedTsConfigPath,
-              ),
-            ]
-          : [],
-      )
-      .filter((reference): reference is string => !!reference);
-
-    for (const referenceConfig of referenceConfigs) {
-      for (const referenceRoot of collectExpandedTsconfigRoots(
-        referenceConfig,
-        config,
-        visited,
-      )) {
-        expandedRoots.add(referenceRoot);
-      }
-    }
-
-    const expandedRootList = [...expandedRoots];
-    tsconfigGraphRootCache.set(tsconfigKey, expandedRootList);
-    debugEmit('expanded tsconfig graph roots', {
-      resolvedTsConfigPath: normalizedTsConfigPath,
-      directRootNameCount: read.rootNames.length,
-      pathRootCount: pathRoots.length,
-      referenceConfigCount: referenceConfigs.length,
-      expandedRootCount: expandedRootList.length,
-    });
-    debugEmitV('expanded tsconfig graph root files', {
-      resolvedTsConfigPath: normalizedTsConfigPath,
-      pathRoots,
-      referenceConfigs,
-      rootNames: expandedRootList,
-    });
-
-    return expandedRootList;
-  }
-
   function resolveTsConfigPath() {
     const tsconfigValue = pluginOptions.tsconfigGetter();
 
@@ -1910,7 +1607,10 @@ export function angular(options?: PluginOptions): Plugin[] {
     }
 
     const resolvedTsConfigPath = resolveTsConfigPath();
-    const cached = getCachedTsconfigOptions(resolvedTsConfigPath, config);
+    const cached = tsconfigResolver.getCachedTsconfigOptions(
+      resolvedTsConfigPath,
+      config,
+    );
 
     // Clone options before mutation (preserve cache purity)
     const tsCompilerOptions = { ...cached.options };
@@ -1958,7 +1658,11 @@ export function angular(options?: PluginOptions): Plugin[] {
       ),
     );
     // Merge + dedupe root names
-    rootNames = union(rootNames, ensureIncludeCache(), replacements);
+    rootNames = union(
+      rootNames,
+      tsconfigResolver.ensureIncludeCache(),
+      replacements,
+    );
     const hostKey = JSON.stringify(tsCompilerOptions);
     let host: ts.CompilerHost;
